fixed keyboard bug (pg 255 dismissing by pressing return key)

+++ BBB/BBB/Item.swift
@@ -23,16 +23,16 @@ class Item: Equatable {
     let dateCreated: Date // not included in draft but should add?
     
     // ratings, will need to make from min 0 to 5, so converts to stars
-    var totalR: Int // average of all stars?
-    var priceR: Int
-    var drinksR: Int
-    var musicR: Int
-    var dancingR: Int
-    var friendlyR: Int
-    let ratings = 5 // all ratings excluding totalR
+    var totalR: Float // average of all stars?
+    var priceR: Float
+    var drinksR: Float
+    var musicR: Float
+    var dancingR: Float
+    var friendlyR: Float
+    let ratings = 5.0 // all ratings excluding totalR
 
     /* INITIALIZER */
-    init(name: String, type: String, address: String, bestFor: String, description: String,  priceR: Int, drinksR: Int, musicR: Int, dancingR: Int, friendlyR: Int){
+    init(name: String, type: String, address: String, bestFor: String, description: String,  priceR: Float, drinksR: Float, musicR: Float, dancingR: Float, friendlyR: Float){
         //basic info
         self.name = name
         self.type = type
@@ -50,7 +50,7 @@ class Item: Equatable {
         self.friendlyR = friendlyR
         
         // get average of ratings
-        self.totalR = (priceR + drinksR + musicR + dancingR + friendlyR ) / ratings
+        self.totalR = (priceR + drinksR + musicR + dancingR + friendlyR ) / Float(ratings)
 
     }
     
@@ -80,7 +80,7 @@ class Item: Equatable {
             // photo
             //let photo = UIImage.init()
             
-            self.init(name: randomName, type: establishmentType, address: randomAddy, bestFor: randomBestFor, description: des,  priceR: Int.random(in:0...5), drinksR: Int.random(in:0...5), musicR: Int.random(in:0...5), dancingR: Int.random(in:0...5), friendlyR: Int.random(in:0...5))
+            self.init(name: randomName, type: establishmentType, address: randomAddy, bestFor: randomBestFor, description: des,  priceR: Float.random(in:0...5), drinksR: Float.random(in:0...5), musicR: Float.random(in:0...5), dancingR: Float.random(in:0...5), friendlyR: Float.random(in:0...5))
             
         } else {
             self.init(name: "", type: "", address: "", bestFor: "", description: "", priceR: 0, drinksR: 0, musicR: 0, dancingR: 0, friendlyR: 0)
